# chore(benchpress): add a ngClass benchmark