--- a/benchmarks/ng-class-bp/bp.conf.js
+++ b/benchmarks/ng-class-bp/bp.conf.js
@@ -1,0 +1,15 @@
+/* eslint-env node */
+
+'use strict';
+
+module.exports = function(config) {
+  config.set({
+    scripts: [{
+      id: 'angular',
+      src: '/build/angular.js'
+    },
+    {
+      src: 'app.js'
+    }]
+  });
+};