**What kind of change does this PR introduce? (Bug fix, feature, docs update, ...)**: benchmark
Adds a benchmark to measure the impact of changes in ngClass.

**What is the current behavior? (You can also link to an open issue here)**: https://github.com/angular/angular.js/pull/15228#commitcomment-19347288 

**What is the new behavior (if this is a feature change)?** no change

**Does this PR introduce a breaking change?** no

**Please check if the PR fulfills these requirements**
- [X] The commit message follows our guidelines: https://github.com/angular/angular.js/blob/master/CONTRIBUTING.md#commit-message-format
- [ ] ~~Tests for the changes have been added (for bug fixes / features)~~
- [ ] ~~Docs have been added / updated (for bug fixes / features)~~

**Other information**:
It adds a benchmark in `benchmarks/ng-class-bp`
The benchmark was designed to measure the changes introduced by: https://github.com/angular/angular.js/pull/15228

It has the following steps:
- _setup_: cleanups preview, and create structures, not worth to measure
- _create_: creation of a table/list with the selected number of items and applied classes
- _$apply_: an apply without changes (compute digest time)
- _update_: cost of updating classes (changes ng-classes values)
- _unclass_: cost of removing all classes (falses all ngClass values)
- _class_: cost of apply classes (applies all classes given already constructed elements)
- _destroy_: cost of removing everything from dom (usually not worth to measure)

There are three versions to paint:
- **optimized table**: a table with optimized ngClass values `ng-class="'success' && todo.completed"`
- **unoptimized table**: a table with object ngClass syntaxis: `ng-class="{success: todo.completed}"`
- **list**: list of todos with an object ngClass and all cases: `ng-class="{completed: todo.completed, important: todo.important, urgent: todo.urgent}"`
- **single optimized**: it computes a single ngClass using a complex object but optimizing the access `ng-class="{success: !!$ctrl.todos, danger: !$ctrl.todos}"`
- **single**: it computes a single ngClass using a complex object `ng-class="{success: $ctrl.todos, danger: !$ctrl.todos}"`

I ran them over the previous version of ngClass and the version of #15228 , the results are:
- create, update, unclass, class:
  - variability in create, update, unclass and class seems to have more or less the same performance
  - the variability seems bigger in the #15228 implementation
- $apply:
  - optimized table:
    - #15228 is around a 25% slower than previous version
    - in my computer previous version took 0.75ms for 1000 rows, and for  #15228 around 1ms
    - I think that this implementation is not often used
  - unoptimized table and list:
    - #15228 is faster than previous version, around 5x and 10x faster
    - I ran the benchmark many times just to be sure
    - previous version has high variability from 7ms ~ 13ms
    -  #15228 maintains the speed around 1ms
  - single optimized:
    - #15228 seems to be slower but there is high variability
    - times are around 0.02ms ~ 0.08ms
  - single:
    - #15228 is really really faster (around more than 100x faster)
    - times of #15228 are close to single optimized
    - previous version takes around 1.5ms

I also modified #15228 to use compile instead of link (commented in https://github.com/angular/angular.js/pull/15228#commitcomment-19360023 ). Results are:
- create:
  - it improves the performance around 5%~10% (faster compile time)
- other cases:
  - no appreciable changes

So, as a conclusion, it seems that the **#15228  optimizes automatically code in the same fashion that an experienced angular programmers would do** manually. 
In the other hand, given that 1 frame is around 16ms, we should give 8ms to the browser to update, it leaves 8ms for the application to work. Experiments show that digest loop times moves from values like 8 or 10ms to just 1ms in large lists, or 0.01ms in a single big object condition.