@drpicox 2016-10-10
I like the result, I usually write guides or check the code in critical zones making sure that my colleagues are using optimized ngClass values. With #15228 seems that it will work well from time 0.

I have rewritted the code a little bit, and fixed the "spelling" problems.

Another question:
- is there any existing tool to execute benchpress from command line? (like karma does with tests)
It would be nice to have benchmarks running automatically and follow the chart of performance changes through the time.